+++ packages/plugin-hardhat/hardhat.config.js
@@ -13,7 +13,7 @@ module.exports = {
   solidity: {
     compilers: [
       {
-        version: '0.8.20',
+        version: '0.8.29',
       },
       {
         version: '0.8.9',